feat(pic): create *pic* module and fixed some clippy issues
