+++ src/arch/ia32/interrupts/pic/pic8259a.rs
@@ -2,43 +2,40 @@
 use crate::arch::io::port::Port;
 use crate::arch::io::register::WriteRegister;
 use crate::utils::bitfield::*;
-use log::trace;
 
 /// Master and Slave PICs ports on IBM PC
-const MASTER_PIC_COM: u16 = 0x20;
-const SLAVE_PIC_COM: u16 = 0xA0;
+pub const MASTER_PIC_COM: u16 = 0x20;
+pub const SLAVE_PIC_COM: u16 = 0xA0;
 
-/// PIC IRQs offset in IDT
-pub const PIC_OFFSET: usize = 32;
 // PIC IRQs mapping
 pub const PIT_IRQ: usize = 0;
 pub const KEYBOARD_IRQ: usize = 1;
 
 /// An enum defining how interrupts are triggered.
-#[derive(PartialEq)]
-enum TriggerMode {
+#[derive(PartialEq, Eq)]
+pub enum TriggerMode {
     /// Level triggered interrupts
-    LEVEL,
+    Level,
     /// Edge triggered interrupts
-    EDGE,
+    Edge,
 }
 
 impl From<TriggerMode> for bool {
     fn from(value: TriggerMode) -> Self {
-        value == TriggerMode::LEVEL
+        value == TriggerMode::Level
     }
 }
 
 /// An enum defining the CALL Address Interval.
-#[derive(PartialEq)]
-enum AddressInterval {
-    FOUR,
-    EIGHT
+#[derive(PartialEq, Eq)]
+pub enum AddressInterval {
+    Four,
+    Eight,
 }
 
 impl From<AddressInterval> for bool {
     fn from(value: AddressInterval) -> Self {
-        value == AddressInterval::FOUR
+        value == AddressInterval::Four
     }
 }
 
@@ -76,15 +73,15 @@ impl From<ICW1> for u8 {
 #[derive(PartialEq)]
 enum PICRole {
     /// No cascading.
-    SINGLE,
+    Single,
     /// Slave in cascading.
-    SLAVE,
+    Slave,
     /// Master in cascading.
-    MASTER,
+    Master,
 }
 
 /// Struct representing a 8259 PIC
-struct PIC8259a {
+pub struct PIC8259a {
     /// Command port of the chip.
     com_port: Port<u8>,
     /// Data port of the chip.
@@ -113,8 +110,8 @@ impl PIC8259a {
             com_port: Port::new(addr),
             data_port: Port::new(addr + 1),
             pin: None,
-            offset: offset,
-            role: PICRole::SINGLE,
+            offset,
+            role: PICRole::Single,
         }
     }
 
@@ -138,8 +135,8 @@ impl PIC8259a {
             com_port: Port::new(addr),
             data_port: Port::new(addr + 1),
             pin: Some(slave_pin),
-            offset: offset,
-            role: PICRole::MASTER,
+            offset,
+            role: PICRole::Master,
         }
     }
 
@@ -163,8 +160,8 @@ impl PIC8259a {
             com_port: Port::new(addr),
             data_port: Port::new(addr + 1),
             pin: Some(slave_pin),
-            offset: offset,
-            role: PICRole::SLAVE,
+            offset,
+            role: PICRole::Slave,
         }
     }
 
@@ -175,24 +172,36 @@ impl PIC8259a {
     /// * `has_icw4` - If true, ICW4 will be expected.
     /// * `adi` - CALL Address Inverval
     /// * `ltim` - Level Triggered mode or Edge Triggered mode.
+    ///
+    /// # Safety
+    ///
+    /// The arguments you are providing should be consistent with other commands you are sending.
     pub unsafe fn send_icw1(&self, has_icw4: bool, adi: AddressInterval, ltim: TriggerMode) {
-        let com = ICW1::new(has_icw4, self.role == PICRole::SINGLE, adi, ltim);
+        let com = ICW1::new(has_icw4, self.role == PICRole::Single, adi, ltim);
         self.com_port.write(com.into());
     }
 
     /// Sends the Initialisation Command Word 2.
+    ///
+    /// # Safety
+    ///
+    /// The arguments you are providing should be consistent with other commands you are sending.
     pub unsafe fn send_icw2(&self) {
         self.data_port.write(self.offset);
     }
 
     /// Sends the Initialisation Command Word 3.
+    ///
+    /// # Safety
+    ///
+    /// The arguments you are providing should be consistent with other commands you are sending.
     pub unsafe fn send_icw3(&self) {
         match self.role {
-            PICRole::SINGLE => panic!("ICW3: Nothing to be sent in single mode."),
-            PICRole::MASTER => self
+            PICRole::Single => panic!("ICW3: Nothing to be sent in single mode."),
+            PICRole::Master => self
                 .data_port
                 .write(0_u8.set_bit(self.pin.unwrap().into(), true)),
-            PICRole::SLAVE => self.data_port.write(self.pin.unwrap()),
+            PICRole::Slave => self.data_port.write(self.pin.unwrap()),
         };
     }
 
@@ -203,11 +212,15 @@ impl PIC8259a {
     /// * `aeoi` - Auto EOI mode.
     /// * `buffered` - Buffered mode.
     /// * `fully_nested` - Special fully nested mode.
+    ///
+    /// # Safety
+    ///
+    /// The arguments you are providing should be consistent with other commands you are sending.
     pub unsafe fn send_icw4(&self, aeoi: bool, buffered: bool, fully_nested: bool) {
         let com = 0_u8
             .set_bit(0, true)
             .set_bit(1, aeoi)
-            .set_bit(2, self.role == PICRole::MASTER)
+            .set_bit(2, self.role == PICRole::Master)
             .set_bit(3, buffered)
             .set_bit(4, fully_nested);
         self.data_port.write(com);
@@ -218,55 +231,20 @@ impl PIC8259a {
     /// # Arguments
     ///
     /// * `mask` - IRQs bitmask (1 = IRQ disabled).
+    ///
+    /// # Safety
+    ///
+    /// The arguments you are providing should be consistent with other commands you are sending.
     pub unsafe fn send_ocw1(&self, mask: u8) {
         self.data_port.write(mask);
     }
 
     /// Acknowledge End Of Interrupt on PIC.
+    ///
+    /// # Safety
+    ///
+    /// Interrupt source should be verified.
     pub unsafe fn ack_eoi(&self) {
         self.com_port.write(0x20_u8);
     }
-}
-
-static MASTER_PIC: PIC8259a = PIC8259a::master(MASTER_PIC_COM, 2, PIC_OFFSET as u8);
-static SLAVE_PIC: PIC8259a = PIC8259a::slave(SLAVE_PIC_COM, 2, (PIC_OFFSET + 8) as u8);
-
-/// Setup both Master and Slave PICs. Sends Initialisation Commands and IRQs masks.
-///
-/// # Arguments
-///
-/// * `master_mask` - Master PIC IRQs bitmask (1 = IRQ disabled).
-/// * `slave_mask` - Slave PIC IRQs bitmask (1 = IRQ disabled).
-pub fn setup(master_mask: u8, slave_mask: u8) {
-    trace!("Setting up 8259a PICs...");
-
-    unsafe {
-        MASTER_PIC.send_icw1(true, AddressInterval::EIGHT, TriggerMode::EDGE);
-        SLAVE_PIC.send_icw1(true, AddressInterval::EIGHT, TriggerMode::EDGE);
-
-        MASTER_PIC.send_icw2();
-        SLAVE_PIC.send_icw2();
-
-        MASTER_PIC.send_icw3();
-        SLAVE_PIC.send_icw3();
-
-        MASTER_PIC.send_icw4(false, false, false);
-        SLAVE_PIC.send_icw4(false, false, false);
-
-        MASTER_PIC.send_ocw1(master_mask);
-        SLAVE_PIC.send_ocw1(slave_mask);
-    }
-}
-
-/// Acknowledge End Of Interrupt on Master PIC or both Master and Slave PICs
-/// depending on the IRQ index.
-///
-/// # Arguments
-///
-/// * `irq` - IRQ index of the EOI to acknowledge.
-pub unsafe fn ack_eoi(irq: u8) {
-    if irq >= 8 {
-        SLAVE_PIC.ack_eoi();
-    }
-    MASTER_PIC.ack_eoi();
 }
